requirements.txt
annotated-types==0.7.0
anyio==4.6.0
bcrypt==4.0.1
cffi==1.17.1
click==8.1.7
fastapi==0.115.0
fastapi-pagination==0.12.28
gunicorn==23.0.0
h11==0.14.0
httptools==0.6.1
idna==3.10
packaging==24.1
passlib==1.7.4
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.12
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.35
sqlmodel==0.0.22
starlette==0.38.6
typing_extensions==4.12.2
uvicorn==0.31.0
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.1
